Refactor Battle.Fight method
To avoid out parameter I refactored Battle.Fight method.
Now it returns tuple.

+++ Seed/Characters/Player.cs
@@ -495,7 +495,7 @@ namespace Seed.Characters
                 Console.WriteLine("umierający.");
 
             Console.WriteLine();
-            var strengthDifference = character.Strength - this.Strength;
+            var strengthDifference = (int)character.Strength - (int)this.Strength;
 
             if(strengthDifference>=10)
                 Console.WriteLine("Nie masz szans w walce.");
